treat all indi gain as ints

+++ indi_allsky/camera/indi.py
@@ -122,7 +122,7 @@ class IndiClient(PyIndi.BaseClient):
         self._timeout = 10.0
 
         self._exposure = 0.0
-        self._gain = 0.0  # individual exposure gain
+        self._gain = 0  # individual exposure gain
 
         self.exposureStartTime = None
 
@@ -206,7 +206,7 @@ class IndiClient(PyIndi.BaseClient):
 
     @gain.setter
     def gain(self, new_gain):
-        self._gain = float(new_gain)
+        self._gain = int(new_gain)  # all indi gain as int
 
 
     @property
@@ -335,7 +335,7 @@ class IndiClient(PyIndi.BaseClient):
         jobdata = {
             'filename'    : str(f_tmpfile_p),
             'exposure'    : self.exposure,
-            'gain'        : self.gain,
+            'gain'        : float(self.gain),
             'exp_time'    : datetime.timestamp(exp_date),  # datetime objects are not json serializable
             'exp_elapsed' : exposure_elapsed_s,
             'camera_id'   : self.camera_id,
@@ -982,7 +982,7 @@ class IndiClient(PyIndi.BaseClient):
 
 
         self.exposure = exposure
-        self.gain = float(self.gain_v.value)
+        self.gain = int(self.gain_v.value)
 
 
         self.exposureStartTime = time.time()
@@ -1024,9 +1024,9 @@ class IndiClient(PyIndi.BaseClient):
 
         # for cameras that do not support gain
         fake_gain_info = {
-            'current' : -1.0,
-            'min'     : -1.0,
-            'max'     : -1.0,
+            'current' : -1,
+            'min'     : -1,
+            'max'     : -1,
             'step'    : 1,
             'format'  : '',
         }
@@ -1084,7 +1084,7 @@ class IndiClient(PyIndi.BaseClient):
                 try:
                     # The label should be the ISO number in string format
                     gain_str = gain_ctl[index].getLabel()
-                    gain = float(gain_str)
+                    gain = int(gain_str)
                 except ValueError:
                     # skip values like "auto"
                     logger.warning('Skipping ISO setting "%s"', gain_str)
@@ -1100,7 +1100,7 @@ class IndiClient(PyIndi.BaseClient):
 
             try:
                 gain_info = {
-                    'current' : 0.0,  # this should not matter
+                    'current' : 0,  # this should not matter
                     'min'     : min(gain_list),
                     'max'     : max(gain_list),
                     'step'    : None,
@@ -1146,7 +1146,7 @@ class IndiClient(PyIndi.BaseClient):
 
 
     def setCcdGain(self, gain_value):
-        logger.warning('Setting CCD gain to %0.1f', float(gain_value))
+        logger.warning('Setting CCD gain to %0.1f', float(int(gain_value)))
         indi_exec = self.ccd_device.getDriverExec()
 
         if indi_exec in [
@@ -1159,7 +1159,7 @@ class IndiClient(PyIndi.BaseClient):
             gain_config = {
                 "PROPERTIES" : {
                     "CCD_CONTROLS" : {
-                        "Gain" : gain_value,
+                        "Gain" : int(gain_value),
                     },
                 },
             }
@@ -1173,7 +1173,7 @@ class IndiClient(PyIndi.BaseClient):
             gain_config = {
                 "PROPERTIES" : {
                     "CCD_GAIN" : {
-                        "GAIN" : gain_value,
+                        "GAIN" : int(gain_value),
                     },
                 },
             }
@@ -1188,7 +1188,7 @@ class IndiClient(PyIndi.BaseClient):
                 gain_config = {
                     "PROPERTIES" : {
                         "CCD_CONTROLS" : {
-                            "Gain" : gain_value,
+                            "Gain" : int(gain_value),
                         },
                     },
                 }
@@ -1197,7 +1197,7 @@ class IndiClient(PyIndi.BaseClient):
                 gain_config = {
                     "PROPERTIES" : {
                         "CCD_GAIN" : {
-                            "GAIN" : gain_value,
+                            "GAIN" : int(gain_value),
                         },
                     },
                 }
@@ -1237,7 +1237,7 @@ class IndiClient(PyIndi.BaseClient):
                 gain_config = {
                     "PROPERTIES" : {
                         "Image Adjustments" : {
-                            "Gain" : gain_value,
+                            "Gain" : int(gain_value),
                         },
                     },
                 }
@@ -1251,7 +1251,7 @@ class IndiClient(PyIndi.BaseClient):
             gain_config = {
                 "PROPERTIES" : {
                     "CCD_GAIN" : {
-                        "GAIN" : gain_value,
+                        "GAIN" : int(gain_value),
                     },
                 },
             }
@@ -1264,7 +1264,7 @@ class IndiClient(PyIndi.BaseClient):
 
         # Update shared gain value
         with self.gain_v.get_lock():
-            self.gain_v.value = int(gain_value)
+            self.gain_v.value = float(int(gain_value))
 
 
     def setCcdBinning(self, bin_value):
@@ -1308,7 +1308,7 @@ class IndiClient(PyIndi.BaseClient):
 
             self.configureDevice(self.ccd_device, binning_config)
 
-            # Update shared gain value
+            # Update shared bin value
             with self.bin_v.get_lock():
                 self.bin_v.value = new_bin_value[0]
         except TimeOutException:
